Training a model to play ZPoker: Zpoker is heads up poker but with constant - changing bet per round. Preflop betting costs 1 token to stay in Flop betting costs 2 tokens to stay in Turn betting costs 3 tokens to stay in River betting costs 4 tokens to stay in
Winning a hand results in a score increase of the above bets.
Losing a hand results in a score decrease of the above bets.

When folding, the player loses all the bets they have made in the round.

This game can be played against a 'house' and a player.

When playing against a player:
A hand is won if the opponent forfeits or the hand is won at showdown.

When playing against the 'house'
The hand can be won at showdown by beating the house, but no other way.